[Add]_03/Initialize_Data
데이터 변수 변경

+++ Assets/Scripts/Eun/Monster_Controller.cs
@@ -7,9 +7,7 @@ public class Monster_Controller : MonoBehaviour  //이동 회전 로직
     [SerializeField]private Monster_Status Datas;
     public float moveDistanse;
 
-
-
-    public int Attack()
+    public float Attack()
     {
         return Datas.attack;
     }
